X00T: init: Drop WiGig from init.qcom.rc
Interface is not in the known set of hidl_interfaces: 'vendor.qti.hardware.wigig.netperftuner@1.0::INetPerfTuner'.
Please ensure the interface is spelled correctly and built by a hidl_interface target.
Interface is not in the known set of hidl_interfaces: 'vendor.qti.hardware.wigig.supptunnel@1.0::ISuppTunnelProvider'.
Please ensure the interface is spelled correctly and built by a hidl_interface target.